refactor(seed): make 51CTO bank the canonical seed; drop demo content
The 51CTO 历年真题 dataset is now the default seed for any self-deploy.
- src/prisma/seed.ts replaces the previous hand-written demo (30 choice +
  5 case mock questions, 2 mock exams) with the 51CTO loader. Running
  npm run db:seed now wipes any non-51CTO Exam/Question records (cascading
  through ExamQuestion, ExamAttempt, options, scenarios, user answers,
  wrong notes) before upserting the 41 papers and full knowledge tree.
- src/prisma/seed-51cto.ts is removed; its logic moved into seed.ts.
- The wipe targets exams whose ID does not start with `exam-51cto-` and
  questions not linked to any 51CTO exam, so a re-seed can't accidentally
  leave orphan demo data behind.

Result on a clean seed: 41 exams, 1579 questions, 96 case scenarios.
